core/src/aura_intelligence/observability/opentelemetry_integration.py
"""
📊 OpenTelemetry Integration - Latest 2025 AI Semantic Conventions
Professional OpenTelemetry integration with AI-specific patterns and latest semantic conventions.

CURRENT STATUS: Temporarily disabled due to import issues
- opentelemetry.semconv.ai module doesn't exist in current version
- Complex dependencies blocking core integration testing
- Using minimal fallback to allow system testing

TODO: Restore full functionality after core integration works
"""

import os
import time
from typing import Dict, Any, Optional
from datetime import datetime, timezone

# ORIGINAL IMPORTS - COMMENTED OUT DUE TO DEPENDENCY ISSUES
# Latest OpenTelemetry imports (2025 patterns)
# from opentelemetry import trace, metrics
# from opentelemetry.sdk.trace import TracerProvider
# from opentelemetry.sdk.trace.export import BatchSpanProcessor
# from opentelemetry.sdk.metrics import MeterProvider
# from opentelemetry.sdk.metrics.export import PeriodicExportingMetricReader
# from opentelemetry.sdk.resources import Resource
# from opentelemetry.exporter.otlp.proto.grpc.trace_exporter import OTLPSpanExporter
# from opentelemetry.exporter.otlp.proto.grpc.metric_exporter import OTLPMetricExporter
# from opentelemetry.instrumentation.auto_instrumentation import sitecustomize
# PROBLEMATIC IMPORT: from opentelemetry.semconv.ai import SpanAttributes as AISpanAttributes
# from opentelemetry.semconv.trace import SpanAttributes
# from opentelemetry.trace import Status, StatusCode

# TEMPORARY MINIMAL IMPORTS FOR TESTING
try:
    from opentelemetry import trace, metrics
    from opentelemetry.trace import Status, StatusCode
    OPENTELEMETRY_AVAILABLE = True
except ImportError:
    OPENTELEMETRY_AVAILABLE = False
    # Create mock objects for testing
    class MockTrace:
        def get_tracer(self, *args, **kwargs): 
            return MockTracer()
    trace = MockTrace()

try:
    from .config import ObservabilityConfig
    from .context_managers import ObservabilityContext
except ImportError:
    # Fallback for direct import
    class ObservabilityConfig:
        def __init__(self, **kwargs):
            self.enabled = kwargs.get('enabled', False)
    
    class ObservabilityContext:
        def __init__(self, **kwargs):
            self.workflow_type = kwargs.get('workflow_type', 'unknown')
            self.workflow_id = kwargs.get('workflow_id', 'unknown')


class MockTracer:
    """Mock tracer for testing when OpenTelemetry not available."""
    def start_span(self, name, **kwargs):
        return MockSpan(name)


class MockSpan:
    """Mock span for testing when OpenTelemetry not available."""
    def __init__(self, name):
        self.name = name
    
    def set_attribute(self, key, value):
        pass
    
    def set_status(self, status):
        pass
    
    def __enter__(self):
        return self
    
    def __exit__(self, *args):
        pass


class OpenTelemetryManager:
    """
    OpenTelemetry integration manager.
    
    CURRENT STATUS: Minimal implementation for testing
    ORIGINAL FEATURES (to be restored):
    - AI-specific span attributes and metrics
    - LLM operation tracking with cost and performance
    - Agent workflow tracing with decision points
    - Automatic instrumentation for popular AI libraries
    - W3C trace context propagation
    - Resource detection and attribution
    """
    
    def __init__(self, config: ObservabilityConfig):
        """Initialize OpenTelemetry manager with fallback support."""
        print("🔧 OpenTelemetryManager: Using minimal implementation for testing")
        self.config = config
        self.enabled = OPENTELEMETRY_AVAILABLE and getattr(config, 'enabled', False)
        
        if self.enabled:
            self.tracer = trace.get_tracer(__name__)
        else:
            self.tracer = MockTracer()
        
        # ORIGINAL ATTRIBUTES (to be restored)
        # self.tracer_provider = None
        # self.meter_provider = None
        # self.meter = None
        # self._active_spans = {}
        # self._llm_token_counter = None
        # self._llm_cost_counter = None
        # self._agent_operation_histogram = None
    
    def start_span(self, name: str, **kwargs):
        """Start a span with fallback support."""
        return self.tracer.start_span(name)
    
    def create_workflow_span(self, context: ObservabilityContext):
        """Create workflow span with minimal attributes."""
        span = self.tracer.start_span(f"workflow.{context.workflow_type}")
        if self.enabled:
            span.set_attribute("workflow.type", context.workflow_type)
        return span
    
    def record_metric(self, name: str, value: float, **kwargs):
        """Record metric with fallback (no-op if not available)."""
        if not self.enabled:
            return
        # TODO: Implement proper metric recording
        pass
    
    async def initialize(self):
        """Initialize with minimal setup."""
        print("🔧 OpenTelemetryManager: Mock initialization complete")
        # TODO: Restore full initialization:
        # - Create AI resource with proper attributes
        # - Initialize tracing with OTLP exporter
        # - Initialize metrics with OTLP exporter
        # - Create metric instruments
        # - Setup auto instrumentation
        pass
    
    async def shutdown(self):
        """Shutdown with minimal cleanup."""
        print("🔧 OpenTelemetryManager: Mock shutdown complete")
        # TODO: Restore full shutdown:
        # - End any remaining spans
        # - Shutdown providers
        pass

    # ORIGINAL METHODS (commented out, to be restored):
    
    # async def start_workflow_span(self, context: ObservabilityContext, state: Dict[str, Any]) -> None:
    #     """Start workflow span with AI semantic conventions."""
    #     pass
    
    # async def complete_workflow_span(self, context: ObservabilityContext, state: Dict[str, Any]) -> None:
    #     """Complete workflow span with results and metrics."""
    #     pass
    
    # async def start_agent_span(self, agent_context: Dict[str, Any]) -> None:
    #     """Start span for agent operation."""
    #     pass
    
    # async def complete_agent_span(self, agent_context: Dict[str, Any]) -> None:
    #     """Complete agent span with results."""
    #     pass
    
    # async def track_llm_usage(self, model_name: str, tokens_used: int, cost_usd: Optional[float] = None) -> None:
    #     """Track LLM usage with AI semantic conventions."""
    #     pass


# Export the main class
__all__ = ['OpenTelemetryManager']

# RESTORATION NOTES:
# 1. Fix opentelemetry.semconv.ai import (use correct semantic conventions)
# 2. Restore full initialization with proper resource creation
# 3. Implement proper span and metric creation
# 4. Add back AI-specific attributes and tracking
# 5. Test with real OpenTelemetry setup